update file names
